sidekickos-client/demos/gemini_live/requirements_gemini.txt
google-genai>=0.8.0
pyaudio>=0.2.11
pillow>=10.0.0 